+++ dartfrog/src/lib.rs
@@ -1,5 +1,4 @@
 use std::collections::{HashMap, HashSet};
-use std::str::FromStr;
 use std::time::{SystemTime, UNIX_EPOCH};
 
 // mod common;
